chore(insights): move trends bar/line charts into product_analytics
First slice of moving product analytics code under products/. Moves the trends bar
and line chart components (along with their shared helpers and handleTrendsChartClick)
from frontend/src/scenes/trends/viz/ into products/product_analytics/frontend/charts/.

Layout:
  products/product_analytics/frontend/charts/
    trends/   -- TrendsBarChart, TrendsLineChart, transforms, bar aggregated click
    shared/   -- AnnotationsLayer, TrendsAlertOverlays, TrendsTooltip, anomaly/goal
                 line adapters, axis formatting, series meta, handleTrendsChartClick

Imports between the moved files use intra-product relative paths; imports back into
still-legacy modules (trendsDataLogic, types, persons-modal, datasetToActorsQuery,
trendsAdapterConstants, scenes/insights/*) use absolute scenes/* aliases so they
remain stable when the surrounding code follows in later slices.

External callers updated:
  - frontend/src/scenes/trends/Trends.tsx lazy imports the moved charts via the
    products/product_analytics/frontend/charts/trends/... path.
  - frontend/src/scenes/funnels/viz/funnel-line-chart/funnelChartTransforms.ts
    imports buildTrendsLineTimeSeriesConfig and buildTrendsSeries from the new
    location.

products/product_analytics/package.json gains posthog-js (used by the moved chart
components) and peer deps for @storybook/react and @testing-library/react (used by
the moved stories and tests).

Generated-By: PostHog Code
Task-Id: 779fb86d-40c5-4855-bb6a-d2f16013824d
