+++ products/product_analytics/frontend/charts/shared/AnnotationsLayer.tsx
