+++ products/product_analytics/frontend/charts/shared/anomalyPointsAdapter.test.ts
@@ -2,7 +2,8 @@ import type { AnomalyPoint } from 'lib/components/Alerts/types'
 
 import { EntityTypes } from '~/types'
 
-import type { IndexedTrendResult } from '../../types'
+import type { IndexedTrendResult } from 'scenes/trends/types'
+
 import { buildAnomalyMarkers } from './anomalyPointsAdapter'
 
 function makeResult(overrides: Partial<IndexedTrendResult> = {}): IndexedTrendResult {
