+++ products/product_analytics/frontend/charts/shared/anomalyPointsAdapter.ts
@@ -1,7 +1,7 @@
 import type { AnomalyPoint } from 'lib/components/Alerts/types'
 import type { AnomalyMarker as BaseAnomalyMarker } from 'lib/hog-charts'
 
-import type { IndexedTrendResult } from '../../types'
+import type { IndexedTrendResult } from 'scenes/trends/types'
 
 export interface AnomalyMarker extends BaseAnomalyMarker {
     /** Originating alert anomaly score (0..1) if available. */
