+++ products/product_analytics/frontend/charts/shared/goalLinesAdapter.test.ts
