+++ products/product_analytics/frontend/charts/shared/handleTrendsChartClick.test.tsx
@@ -1,7 +1,8 @@
 import { NodeKind } from '~/queries/schema/schema-general'
 import { CompareLabelType, EntityTypes } from '~/types'
 
-import type { IndexedTrendResult } from '../types'
+import type { IndexedTrendResult } from 'scenes/trends/types'
+
 import { handleTrendsChartClick, type TrendsChartClickDeps } from './handleTrendsChartClick'
 
 function makeTrendResult(overrides: Partial<IndexedTrendResult> = {}): IndexedTrendResult {
