+++ products/product_analytics/frontend/charts/shared/trendsAxisFormat.ts
