+++ products/product_analytics/frontend/charts/shared/trendsSeriesMeta.ts
