+++ products/product_analytics/frontend/charts/trends/TrendsBarChart.stories.tsx
