+++ products/product_analytics/frontend/charts/trends/TrendsBarChart.test.tsx
