+++ products/product_analytics/frontend/charts/trends/TrendsBarChart.tsx
@@ -23,12 +23,13 @@ import { InsightVizNode } from '~/queries/schema/schema-general'
 import { QueryContext } from '~/queries/types'
 import { ChartDisplayType } from '~/types'
 
-import { openPersonsModal } from '../../persons-modal/PersonsModal'
-import { trendsDataLogic } from '../../trendsDataLogic'
-import type { IndexedTrendResult } from '../../types'
-import { handleTrendsChartClick, type TrendsChartClickDeps } from '../handleTrendsChartClick'
+import { openPersonsModal } from 'scenes/trends/persons-modal/PersonsModal'
+import { trendsDataLogic } from 'scenes/trends/trendsDataLogic'
+import type { IndexedTrendResult } from 'scenes/trends/types'
+
 import { AnnotationsLayer } from '../shared/AnnotationsLayer'
 import { goalLinesToReferenceLines } from '../shared/goalLinesAdapter'
+import { handleTrendsChartClick, type TrendsChartClickDeps } from '../shared/handleTrendsChartClick'
 import { TrendsAlertOverlays } from '../shared/TrendsAlertOverlays'
 import { trendsFilterToYFormatterConfig } from '../shared/trendsAxisFormat'
 import type { TrendsSeriesMeta } from '../shared/trendsSeriesMeta'
