+++ products/product_analytics/frontend/charts/trends/TrendsLineChart.tsx
@@ -15,12 +15,13 @@ import { groupsModel } from '~/models/groupsModel'
 import { InsightVizNode } from '~/queries/schema/schema-general'
 import { QueryContext } from '~/queries/types'
 
-import { InsightEmptyState } from '../../../insights/EmptyStates'
-import { openPersonsModal } from '../../persons-modal/PersonsModal'
-import { trendsDataLogic } from '../../trendsDataLogic'
-import type { IndexedTrendResult } from '../../types'
-import { handleTrendsChartClick } from '../handleTrendsChartClick'
+import { InsightEmptyState } from 'scenes/insights/EmptyStates'
+import { openPersonsModal } from 'scenes/trends/persons-modal/PersonsModal'
+import { trendsDataLogic } from 'scenes/trends/trendsDataLogic'
+import type { IndexedTrendResult } from 'scenes/trends/types'
+
 import { AnnotationsLayer } from '../shared/AnnotationsLayer'
+import { handleTrendsChartClick } from '../shared/handleTrendsChartClick'
 import { TrendsAlertOverlays } from '../shared/TrendsAlertOverlays'
 import type { TrendsSeriesMeta } from '../shared/trendsSeriesMeta'
 import { TrendsTooltip } from '../shared/TrendsTooltip'
