+++ products/product_analytics/frontend/charts/trends/handleTrendsBarAggregatedChartClick.test.ts
@@ -1,8 +1,9 @@
 import { NodeKind } from '~/queries/schema/schema-general'
 import { EntityTypes } from '~/types'
 
-import type { IndexedTrendResult } from '../../types'
-import type { TrendsChartClickDeps } from '../handleTrendsChartClick'
+import type { IndexedTrendResult } from 'scenes/trends/types'
+
+import type { TrendsChartClickDeps } from '../shared/handleTrendsChartClick'
 import { handleTrendsBarAggregatedChartClick } from './handleTrendsBarAggregatedChartClick'
 
 function makeTrendResult(overrides: Partial<IndexedTrendResult> = {}): IndexedTrendResult {
