+++ products/product_analytics/frontend/charts/trends/trendsBarChartTransforms.test.ts
