+++ products/product_analytics/frontend/charts/trends/trendsChartTransforms.test.ts
